ai/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.2.0
faiss-cpu>=1.8.0
openai>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
